ecoshop-main/reward_orchestrator/requirements.txt
flask
pika
requests
python-dotenv